[QEMU][RFC] hw/arm/sysbus-fdt: Enable rcar-gen3-scif dynamic instantiation
Allow the provision of a Renesas R-Car Gen3 (H)SCIF controller device
from the Qemu command line:

  -device vfio-platform,host=e6e68000.serial,manufacturer=renesas,model=rcar-gen3-scif

To support this, the serial device must be unbound from the host, and
rebound to the vfio-platform driver:

  echo 1 > /sys/module/vfio/parameters/enable_unsafe_noiommu_mode
  echo e6e68000.serial > /sys/bus/platform/drivers/sh-sci/unbind
  echo vfio-platform > /sys/bus/platform/devices/e6e68000.serial/driver_override
  echo e6e68000.serial > /sys/bus/platform/drivers/vfio-platform/bind

Ideally, this patch should be unnecessary and the FDT should be
generated automatically from the host, or perhaps with the provision of
a DTO, allowing the configuration to be provided externally from Q-Emu.

The functionality of this patch mostly duplicates the
add_rcar_gpio_fdt_node() with the exception of the gpio-controller node,
and #gpio-cells.

Not-Signed-off-by: Kieran Bingham <kieran.bingham+renesas@ideasonboard.com>

---

This patch and its dependencies is available at the following branch:

  https://github.com/kbingham/qemu.git rcar3/serial/passthrough